feature: build basic layout putting the necessary composable elements
Include in the composable functions the necessary elements that have to be displayed in the UI.
These are the artwork, with the image, title, artist and year, and the button row, where the previous and next buttons are added.
Add the necessary string resources in the strings.xml resource file.

Integrates #1