feature: refine layout and improve the user interface
Improve the user interface of the composable layout. Specifically, introduce spacings between the image and the text in the artwork presentation, then put the buttons to the bottom. Add then a surface around the picture.
Shift the logic for composing the artwork in the ArtworkPresentation composable.
Remove the id variable inside the Artwork data class.

Integrates #1